moved transform search suggestion to its own function
